feat: custom domain backend

+++ apps/api-journeys/src/app/modules/customDomain/customDomain.resolver.ts
@@ -1,3 +1,4 @@
+import { subject } from '@casl/ability'
 import { UseGuards } from '@nestjs/common'
 import {
   Args,
@@ -8,13 +9,17 @@ import {
   ResolveReference,
   Resolver
 } from '@nestjs/graphql'
+import omit from 'lodash/omit'
 
 import { CustomDomain } from '.prisma/api-journeys-client'
+import { CaslAbility } from '@core/nest/common/CaslAuthModule'
 
 import {
   CustomDomainCreateInput,
+  CustomDomainUpdateInput,
   CustomDomainVerification
 } from '../../__generated__/graphql'
+import { Action, AppAbility } from '../../lib/casl/caslFactory'
 import { AppCaslGuard } from '../../lib/casl/caslGuard'
 import { PrismaService } from '../../lib/prisma.service'
 
@@ -46,15 +51,56 @@ export class CustomDomainResolver {
   @Mutation()
   @UseGuards(AppCaslGuard)
   async customDomainCreate(
-    @Args('input') input: CustomDomainCreateInput
+    @Args('input') input: CustomDomainCreateInput,
+    @CaslAbility() ability: AppAbility
   ): Promise<CustomDomain & { verification: CustomDomainVerification }> {
-    return await this.customDomainService.customDomainCreate(input)
+    return await this.customDomainService.customDomainCreate(input, ability)
   }
 
   @Mutation()
   @UseGuards(AppCaslGuard)
-  async customDomainDelete(@Args('id') id: string): Promise<CustomDomain> {
-    const customDomain = await this.prismaService.customDomain.delete({
+  async customDomainUpdate(
+    @Args('input') input: CustomDomainUpdateInput,
+    @CaslAbility() ability: AppAbility
+  ): Promise<CustomDomainUpdateInput> {
+    const customDomain = await this.prismaService.customDomain.findUnique({
+      where: { id: input.id }
+    })
+    if (customDomain == null) {
+      throw new Error('Custom domain not found')
+    }
+    if (!ability.can(Action.Manage, subject('CustomDomain', customDomain))) {
+      throw new Error('user is not allowed to update custom domain')
+    }
+    return await this.prismaService.customDomain.update({
+      where: { id: input.id },
+      data: {
+        ...omit(input, 'journeyCollectionId'),
+        name: input.name ?? undefined,
+        allowOutsideJourneys: input.allowOutsideJourneys ?? undefined,
+        journeyCollection: {
+          connect: { id: input.journeyCollectionId ?? undefined }
+        }
+      }
+    })
+  }
+
+  @Mutation()
+  @UseGuards(AppCaslGuard)
+  async customDomainDelete(
+    @Args('id') id: string,
+    @CaslAbility() ability: AppAbility
+  ): Promise<CustomDomain> {
+    const customDomain = await this.prismaService.customDomain.findUnique({
+      where: { id }
+    })
+    if (customDomain == null) {
+      throw new Error('Custom domain not found')
+    }
+    if (!ability.can(Action.Delete, subject('CustomDomain', customDomain))) {
+      throw new Error('user is not allowed to delete custom domain')
+    }
+    await this.prismaService.customDomain.delete({
       where: { id }
     })
     await this.customDomainService.deleteVercelDomain(customDomain.name)
